feat(logger): introduce rich logging with loguru for enhanced output
Replace the existing logging setup with a new logger module
utilizing loguru and rich for improved log formatting and
readability. This change enhances the logging experience by
providing colored level indicators, timestamps, and source
location information. It also supports splitting long messages
into multiple lines for better clarity.

fix(tty_roles): remove unnecessary int conversion for exponent

The conversion of the exponent to an integer is redundant as
math.floor already returns an integer. This change simplifies
the code.

refactor(help): use pathlib for directory listing in _get_cog_groups

Replace os.listdir with pathlib's iterdir for improved readability
and consistency with other parts of the codebase. This change
also removes the unnecessary import of os.

chore(mise.toml): update dependencies and add node

Update the Python version to 3.13.2 for compatibility and add
node as a tool dependency to ensure the latest version is used
for any node-related tasks.

+++ tux/cogs/services/tty_roles.py
@@ -53,7 +53,7 @@ class TtyRoles(commands.Cog):
         if 1 <= user_count <= 128:
             return f"{self.base_role_name}0"
 
-        exponent = int(math.floor(math.log2(user_count)))
+        exponent = math.floor(math.log2(user_count))
 
         return f"{self.base_role_name}{2**exponent}"
